add terraforming assets add nice visual
