+++ Resources/Textures/Objects/Decoration/Flora/flora_shadow_trees.rsi/meta.json
@@ -1,0 +1,29 @@
+{
+    "version": 1,
+    "license": "CC-BY-SA-3.0",
+    "copyright": "Taken from tgstation at commit https://github.com/tgstation/tgstation/commit/e00cae8d065f9cf520688cc0dd0e15ba5bef12a9 and recolor by TheShuEd",
+    "size": {
+        "x": 96,
+        "y": 96
+    },
+    "states": [
+        {
+            "name": "tree01"
+        },
+        {
+            "name": "tree02"
+        },
+        {
+            "name": "tree03"
+        },
+        {
+            "name": "tree04"
+        },
+        {
+            "name": "tree05"
+        },
+        {
+            "name": "tree06"
+        }
+    ]
+}
